+++ osprey/execute_skeleton.py
@@ -6,7 +6,7 @@ from pkg_resources import resource_filename
 TEMPLATES = {'msmbuilder': 'msmbuilder_skeleton_config.yaml',
              'sklearn': 'sklearn_skeleton_config.yaml',
              'random_example': 'random_example.yaml',
-             'gp_example': 'sklearn_skeleton_config.yaml',
+             'bayes_example': 'sklearn_skeleton_config.yaml',
              'grid_example': 'grid_example.yaml',
              'msmb_feat_select': 'msmb_feat_select_skeleton_config.yaml'}