+++ osprey/strategies.py
@@ -13,20 +13,10 @@ except ImportError:
     # hyperopt is optional, but required for hyperopt_tpe()
     pass
 
-try:
-    from GPy import kern
-    from GPy.kern import RBF, Fixed, Bias
-    from GPy.util.linalg import tdot
-    from GPy.models import GPRegression
-    from scipy.optimize import minimize
-    from scipy.stats import norm
-    # If the GPy modules fail we won't do this unnecessarily.
-    KERNEL_BASE_CLASS = kern.src.kern.Kern
-except ImportError:
-    # GPy is optional, but required for gp
-    GPRegression = kern = minimize = None
-    pass
 from .search_space import EnumVariable
+from .acquisition_functions import AcquisitionFunction
+from .surrogate_models import (MaximumLikelihoodGaussianProcess,
+                               GaussianProcessKernel)
 
 DEFAULT_TIMEOUT = socket._GLOBAL_DEFAULT_TIMEOUT
 
@@ -92,8 +82,10 @@ class RandomSearch(BaseStrategy):
 class HyperoptTPE(BaseStrategy):
     short_name = 'hyperopt_tpe'
 
-    def __init__(self, seed=None):
+    def __init__(self, seed=None, gamma=0.25, seeds=20):
         self.seed = seed
+        self.gamma = gamma
+        self.seeds = seeds
 
     def suggest(self, history, searchspace):
         """
@@ -183,6 +175,12 @@ class HyperoptTPE(BaseStrategy):
         trials.refresh()
         chosen_params_container = []
 
+        def suggest(*args, **kwargs):
+            return tpe.suggest(*args,
+                               **kwargs,
+                               gamma=self.gamma,
+                               n_startup_jobs=self.seeds)
+
         def mock_fn(x):
             # http://stackoverflow.com/a/3190783/1079728
             # to get around no nonlocal keywork in python2
@@ -210,121 +208,41 @@ class HyperoptTPE(BaseStrategy):
         return kwargs
 
 
-class GP(BaseStrategy):
-    short_name = 'gp'
+class Bayes(BaseStrategy):
+    short_name = 'bayes'
 
     def __init__(self,
                  acquisition=None,
+                 surrogate=None,
+                 kernels=None,
                  seed=None,
                  seeds=1,
                  max_feval=5E4,
-                 max_iter=1E5):
+                 max_iter=1E5,
+                 n_iter=50):
         self.seed = seed
         self.seeds = seeds
         self.max_feval = max_feval
         self.max_iter = max_iter
-        self.model = None
+        self.n_iter = n_iter
         self.n_dims = None
-        self.kernel = None
-        self._kerns = None
-        self._kernf = None
-        self._kernb = None
+        if surrogate is None:
+            surrogate = 'gp'
+        self.surrogate = surrogate
+        if kernels is None:
+            kernels = [{
+                'name': 'GPy.kern.Matern52',
+                'params': {
+                    'ARD': True
+                },
+                'options': {
+                    'independent': False
+                }
+            }]
+        self.kernel_params = kernels
         if acquisition is None:
             acquisition = {'name': 'osprey', 'params': {}}
-        self.acquisition_function = acquisition
-        self._acquisition_function = None
-        self._set_acquisition()
-
-    def _create_kernel(self, V):
-        self._kerns = [
-            RBF(1, ARD=True, active_dims=[i]) for i in range(self.n_dims)
-        ]
-        self._kernf = Fixed(self.n_dims, tdot(V))
-        self._kernb = Bias(self.n_dims)
-        self.kernel = np.sum(self._kerns) + self._kernf + self._kernb
-
-    def _fit_model(self, X, Y):
-        model = GPRegression(X, Y, self.kernel)
-        model.optimize(messages=False, max_f_eval=self.max_feval)
-        self.model = model
-
-    def _get_random_point(self):
-        return np.array(
-            [np.random.uniform(low=0., high=1.) for i in range(self.n_dims)])
-
-    def _is_var_positive(self, var):
-
-        if np.any(var < 0):
-            # RuntimeError may be overkill
-            raise RuntimeError(
-                'Negative variance predicted from regression model.')
-        else:
-            return True
-
-    def _ei(self, x, y_mean, y_var):
-        y_std = np.sqrt(y_var)
-        y_best = self.model.Y.max(axis=0)
-        z = (y_mean - y_best) / y_std
-        result = y_std * (z * norm.cdf(z) + norm.pdf(z))
-        return result
-
-    def _ucb(self, x, y_mean, y_var, kappa=1.0):
-        result = y_mean + kappa * np.sqrt(y_var)
-        return result
-
-    def _osprey(self, x, y_mean, y_var):
-        return (y_mean + y_var).flatten()
-
-    def _optimize(self, init=None):
-        # TODO start minimization from a range of points and take minimum
-        if not init:
-            init = self._get_random_point()
-
-        def z(x):
-            # TODO make spread of points around x and take mean value.
-            x = x.copy().reshape(-1, self.n_dims)
-            y_mean, y_var = self.model.predict(
-                x, kern=(np.sum(self._kerns).copy() + self._kernb.copy()))
-            # This code is for debug/testing phase only.
-            # Ideally we should test for negative variance regardless of the AF.
-            # However, we want to recover the original functionality of Osprey, hence the conditional block.
-            # TODO remove this.
-            if self.acquisition_function['name'] == 'osprey':
-                af = self._acquisition_function(x, y_mean=y_mean, y_var=y_var)
-            elif self.acquisition_function['name'] in ['ei', 'ucb']:
-                # y_var = np.abs(y_var)
-                if self._is_var_positive(y_var):
-                    af = self._acquisition_function(x,
-                                                    y_mean=y_mean,
-                                                    y_var=y_var)
-            return (-1) * af
-
-        res = minimize(z,
-                       init,
-                       bounds=self.n_dims * [(0., 1.)],
-                       options={
-                           'maxiter': self.max_iter,
-                           'disp': 0
-                       })
-        return res.x
-
-    def _set_acquisition(self):
-        if isinstance(self.acquisition_function, list):
-            raise RuntimeError('Must specify only one acquisition function')
-        if sorted(self.acquisition_function.keys()) != ['name', 'params']:
-            raise RuntimeError('strategy/params/acquisition must contain keys '
-                               '"name" and "params"')
-        if self.acquisition_function['name'] not in ['ei', 'ucb', 'osprey']:
-            raise RuntimeError(
-                'strategy/params/acquisition name must be one of '
-                '"ei", "ucb", "osprey"')
-
-        f = eval('self._' + self.acquisition_function['name'])
-
-        def g(x, y_mean, y_var):
-            return f(x, y_mean, y_var, **self.acquisition_function['params'])
-
-        self._acquisition_function = g
+        self.acquisition_params = acquisition
 
     def _get_data(self, history, searchspace):
         X = []
@@ -338,7 +256,7 @@ class GP(BaseStrategy):
                 # not sure how to deal with these yet
                 continue
 
-            point = searchspace.point_to_gp(param_dict)
+            point = searchspace.point_to_unit(param_dict)
             if status == 'SUCCEEDED':
                 X.append(point)
                 Y.append(np.mean(scores))
@@ -353,7 +271,7 @@ class GP(BaseStrategy):
                 np.array(Y).reshape(-1, 1), np.array(V).reshape(-1, 1),
                 np.array(ignore).reshape(-1, self.n_dims))
 
-    def _from_gp(self, result, searchspace):
+    def _from_unit(self, result, searchspace):
 
         # Note that GP only deals with float-valued variables, so we have
         # a transform step on either side, where int and enum valued variables
@@ -361,8 +279,7 @@ class GP(BaseStrategy):
         # GP needs to be reverse-transformed.
         out = {}
         for gpvalue, var in zip(result, searchspace):
-            out[var.name] = var.point_from_gp(float(gpvalue))
-
+            out[var.name] = var.point_from_unit(float(gpvalue))
         return out
 
     def _is_within(self, point, X, tol=1E-2):
@@ -371,27 +288,37 @@ class GP(BaseStrategy):
         return False
 
     def suggest(self, history, searchspace, max_tries=5):
-        if not GPRegression:
-            raise ImportError('No module named GPy')
-        if not minimize:
-            raise ImportError('No module named SciPy')
-
         if len(history) < self.seeds:
             return RandomSearch().suggest(history, searchspace)
 
         self.n_dims = searchspace.n_dims
 
         X, Y, V, ignore = self._get_data(history, searchspace)
-
         # TODO make _create_kernel accept optional args.
-        self._create_kernel(V)
-        self._fit_model(X, Y)
-        suggestion = self._optimize()
+        # Define and fit model
+        if self.surrogate == 'gp':
+            kernel = GaussianProcessKernel(self.kernel_params, self.n_dims)
+            model = MaximumLikelihoodGaussianProcess(X=X,
+                                                     Y=Y,
+                                                     kernel=kernel.kernel,
+                                                     max_feval=self.max_feval)
+        else:
+            raise NotImplementedError(
+                'Surrogate model not recognised.  Please choose from: gp')
+        model.fit()
+
+        # Define acquisition function and get best candidate
+        af = AcquisitionFunction(surrogate=model,
+                                 acquisition_params=self.acquisition_params,
+                                 n_dims=self.n_dims,
+                                 n_iter=self.n_iter,
+                                 max_iter=self.max_iter)
+        suggestion = af.get_best_candidate()
 
         if suggestion in ignore or self._is_within(suggestion, X):
             return RandomSearch().suggest(history, searchspace)
 
-        return self._from_gp(suggestion, searchspace)
+        return self._from_unit(suggestion, searchspace)
 
 
 class GridSearch(BaseStrategy):